web-server/app/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.2
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
chardet==5.2.0
click==8.3.0
distro==1.9.0
docx2txt==0.9
EbookLib==0.19
Flask==3.1.2
Flask-Caching==1.10.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
lxml==6.0.2
MarkupSafe==3.0.3
openai==2.1.0
pillow
pydantic==2.11.10
pydantic_core==2.33.2
PyMuPDF==1.26.4
PyPDF2==1.24
python-docx==1.2.0
python-pptx==1.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
textract==1.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
xlrd==2.0.2
xlsxwriter==3.2.9
redis
cachelib
uuid